backend-java/mm-clip/python_clip/clip_inference.py
import torch
import clip
from PIL import Image

class CLIPModel:
    def __init__(self):
        self.device = "cuda" if torch.cuda.is_available() else "cpu"
        self.model, self.preprocess = clip.load("ViT-B/32", device=self.device)

    def calculate_similarity(self, image_path, text):
        image = self.preprocess(Image.open(image_path)).unsqueeze(0).to(self.device)
        text = clip.tokenize([text]).to(self.device)

        with torch.no_grad():
            image_features = self.model.encode_image(image)
            text_features = self.model.encode_text(text)
            similarity = torch.cosine_similarity(image_features, text_features)
        
        return similarity.item()
