backend-java/mm-clip/python_clip/requirements.txt
torch
clip
flask
Pillow
